@@ -2,7 +2,8 @@ package io.donsaad.movies247.movies;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import io.donsaad.movies247.R;
 import io.donsaad.movies247.moviedetails.DetailsActivity;
